🚚 Move openmc to codes
